+++ src/controllers/OrdersController.php
@@ -17,6 +17,7 @@ use craft\commerce\elements\Order;
 use craft\commerce\errors\CurrencyException;
 use craft\commerce\errors\RefundException;
 use craft\commerce\errors\TransactionException;
+use craft\commerce\events\OrderTabsEvent;
 use craft\commerce\gateways\MissingGateway;
 use craft\commerce\helpers\Purchasable;
 use craft\commerce\models\Address;
@@ -63,6 +64,31 @@ use yii\web\Response;
  */
 class OrdersController extends Controller
 {
+
+
+    /**
+     * @event OrderTabsEvent The event allows additional tabs to be added or modified on the Edit Order page.
+     * @since 3.2.9
+     *
+     * ```php
+     * Event::on(
+     * OrdersController::class,
+     * OrdersController::EVENT_REGISTER_ORDER_TABS,
+     *     function(OrderTabsEvent $event) {
+     *         $tabs = $event->tabs;
+     *         $tabs[] = [
+     *           'label' => Craft::t('commerce', 'Mine'),
+     *           'url' => '#mineTab',
+     *           'class' => null
+     *         ];
+     *         $event->tabs = $tabs;
+     *     }
+     * );
+     * ```
+     */
+    const EVENT_REGISTER_ORDER_TABS = 'beforeSaveCustomer';
+
+
     /**
      * @throws HttpException
      * @throws InvalidConfigException
@@ -927,6 +953,15 @@ class OrdersController extends Controller
             'url' => '#orderHistoryTab',
             'class' => null
         ];
+
+        $orderTabsEvent = new OrderTabsEvent();
+        $orderTabsEvent->tabs = $variables['tabs'];
+
+        if ($this->hasEventHandlers(self::EVENT_REGISTER_ORDER_TABS)) {
+            $this->trigger(self::EVENT_REGISTER_ORDER_TABS, $orderTabsEvent);
+        }
+
+        $variables['tabs'] = $orderTabsEvent->tabs;
 
         $variables['fullPageForm'] = true;